influxdb_connector/requirements.txt
certifi==2022.6.15
influxdb-client==1.30.0
kafka-python==2.0.2
python-dateutil==2.8.2
Rx==3.2.0
six==1.16.0
urllib3==1.26.10
